compiler: refactor handing of functions
Previously, all functions were processed first by the ir package which
scanned them all for reachability. Only the reachable functions were
compiled to LLVM IR.

This change removes this dead code elimination pass, and in fact
refactors most of that infrastructure. This cleans up the compiler and
should open the door for caching the compilation and optimization of
packages.